requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.7.0
django-ninja==1.3.0
django-ninja-extra==0.22.3
django-ninja-jwt==5.3.5
gunicorn==23.0.0
injector==0.22.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
packaging==24.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
